requirements.txt
appdirs==1.4.3
asgi-redis==1.3.0
asgiref==1.1.1
attrs==16.3.0
autobahn==0.18.2
Automat==0.5.0
channels==1.1.3
constantly==15.1.0
daphne==1.2.0
dj-database-url==0.4.1
Django==1.10.3
django-braces==1.8.1
django-cors-headers==2.0.2
django-oauth-toolkit==0.10.0
django-rest-auth==0.7.0
djangorestframework==3.6.2
djangorestframework-jwt==1.9.0
dropbox==7.2.1
gunicorn==19.6.0
incremental==16.10.1
msgpack-python==0.4.8
oauthlib==1.0.3
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
pypiwin32==219
python-dateutil==2.5.3
python-mimeparse==1.5.2
redis==2.10.5
requests==2.13.0
six==1.10.0
Twisted==17.1.0
txaio==2.7.0
typing==3.6.1
urllib3==1.20
whitenoise==3.2.1
zope.interface==4.4.0
